+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.0.3] - 11/05/2021.
+
+* Support 12 hours format
+* Add more style customizable
+* Editable time value
+* Provide documentation
+
 ## [0.0.2] - 22/04/2021.
 
 * Add initial value param